# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "LedgerLine backend API.",
        "title": "LedgerLine API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Tim LedgerLine",
            "email": "dev@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/roles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "List roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "-is_system,name",
                        "description": "Order: name, is_system, user_count, created_at, updated_at. Prefix - for desc",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RoleResponseDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Create a new role",
                "parameters": [
                    {
                        "description": "Role data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.CreateRoleRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RoleResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            }
        },
        "/roles/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Role detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RoleResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Delete a role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Update role data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.UpdateRoleRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RoleResponseDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.CreateRoleMenuPermissionRequestDTO": {
            "type": "object",
            "required": [
                "menu_id"
            ],
            "properties": {
                "can_approve": {
                    "type": "boolean",
                    "example": true
                },
                "can_create": {
                    "type": "boolean",
                    "example": true
                },
                "can_delete": {
                    "type": "boolean",
                    "example": true
                },
                "can_read": {
                    "type": "boolean",
                    "example": true
                },
                "can_update": {
                    "type": "boolean",
                    "example": true
                },
                "menu_id": {
                    "type": "string",
                    "example": "b0000000-0000-0000-0000-000000000002"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.CreateRoleRequestDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Handles day to day transactions"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "Finance Staff"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.CreateRoleMenuPermissionRequestDTO"
                    }
                }
            }
        },
...
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RoleMenuPermissionResponseDTO": {
            "type": "object",
            "properties": {
                "can_approve": {
                    "type": "boolean",
                    "example": true
                },
                "can_create": {
                    "type": "boolean",
                    "example": true
                },
                "can_delete": {
                    "type": "boolean",
                    "example": true
                },
                "can_read": {
                    "type": "boolean",
                    "example": true
                },
                "can_update": {
                    "type": "boolean",
                    "example": true
                },
                "menu_id": {
                    "type": "string",
                    "example": "b0000000-0000-0000-0000-000000000002"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RoleResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-01-02T15:04:05Z"
                },
                "description": {
                    "type": "string",
                    "example": "Built-in role with access to every menu"
                },
                "id": {
                    "type": "string",
                    "example": "00000000-0000-0000-0000-000000000001"
                },
                "is_system": {
                    "description": "Built-in roles must not be renamed or deleted; the UI hides those actions.",
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Admin"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.RoleMenuPermissionResponseDTO"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-01-02T15:04:05Z"
                },
                "user_count": {
                    "description": "How many users are assigned to this role; filled by the list only.",
                    "type": "integer",
                    "example": 12
                }
            }
        },
...
        "github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.UpdateRoleRequestDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Approves transactions above the limit"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "Finance Lead"
                },
                "permissions": {
                    "description": "Omitted leaves the permissions untouched; an empty array clears them.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_internal_delivery_http_dto.CreateRoleMenuPermissionRequestDTO"
                    }
                }
            }
        },
...
        "github_com_stevensuki_ledgerline-backend_pkg_response.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "USER_NOT_FOUND"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.FieldError"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "something went wrong"
                },
                "request_id": {
                    "type": "string",
                    "example": "9b2c1f2e-6f0a-4a1e-9c7d-2f8b0a1c3d4e"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_pkg_response.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "email"
                },
                "message": {
                    "type": "string",
                    "example": "email must be a valid email address"
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "per_page": {
                    "type": "integer",
                    "example": 10
                },
                "total_items": {
                    "type": "integer",
                    "example": 42
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "github_com_stevensuki_ledgerline-backend_pkg_response.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_stevensuki_ledgerline-backend_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Fill with: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}